pyproject.toml
[project]
name = "depths"
version = "0.1.0"
description = "Data Layer for AI"
readme = "README.md"
authors = [
    { name = "Depths AI", email = "admin@depthsai.com" }
]
requires-python = ">=3.12"
dependencies = [
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "openai>=1.98.0",
    "polars==1.31.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
depths = "depths.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["depths/**/*.py", "README.md"]

[tool.hatch.build.targets.wheel]
packages=["depths"]
